requirements.txt
streamlit
transformers
torch
sentencepiece
TTS
openai-whisper
pydub